Move event class to a new file
